requirements.txt
streamlit
streamlit_card
google-generativeai
python-dotenv
newsapi-python
xmltodict
